test: KingTest 코드 작성
